refactor: move api models to subfolder in API
