+++ src/depot_server/api/device.py
@@ -3,11 +3,11 @@ from typing import List, Tuple
 from authlib.oidc.core import UserInfo
 from fastapi import APIRouter, Depends, Body, BackgroundTasks, Response
 
-from depot_server.api.models.reservations import reservation_action_impl
+from depot_server.api.models import Reservation, DeviceReservation, Item, Bay, ItemCondition, User, \
+    ReservationActionInWrite
+from depot_server.api.reservations import reservation_action_impl
 from depot_server.db import collections, DbReservation, DbItem, DbItemReservation
 from depot_server.helper.auth import DeviceAuthentication
-from depot_server.model import Reservation, DeviceReservation, Item, Bay, ItemCondition, User, \
-    ReservationActionInWrite
 
 router = APIRouter()
