+++ src/depot_server/api/item_history.py
@@ -6,9 +6,9 @@ from authlib.oidc.core import UserInfo
 from fastapi import APIRouter, Depends, Query, HTTPException
 from pymongo import DESCENDING, ASCENDING
 
+from depot_server.api.models import ItemState
 from depot_server.db import collections
 from depot_server.helper.auth import Authentication
-from depot_server.model import ItemState
 
 router = APIRouter()
