+++ src/depot_server/api/models/base.py
