+++ src/depot_server/api/models/device_reservation.py
