+++ src/depot_server/api/models/item_state.py
@@ -1,15 +1,14 @@
-from enum import Enum
-
 from datetime import datetime
-from pydantic import Field
+from enum import Enum
 from typing import List, Optional
 from uuid import UUID
 
+from pydantic import Field
+
+from depot_server.helper.util import date
 from .base import BaseModel
 from .report_element import ReportState
 from .report_profile import TotalReportState
-from depot_server.helper.util import date
-
 
 
 class ItemCondition(str, Enum):
