+++ src/depot_server/api/models/picture.py
