+++ src/depot_server/api/models/report_element.py
@@ -1,7 +1,7 @@
 from enum import Enum
+from uuid import UUID
 
 from pydantic import Field
-from uuid import UUID
 
 from .base import BaseModel
