+++ src/depot_server/api/users.py
@@ -3,8 +3,8 @@ from typing import List
 from authlib.oidc.core import UserInfo
 from fastapi import APIRouter, Depends, HTTPException
 
+from depot_server.api.models import User
 from depot_server.helper.auth import Authentication, get_profiles, get_profile
-from depot_server.model import User
 
 router = APIRouter()
