+++ src/depot_server/db/model/item_state.py
@@ -1,13 +1,13 @@
 from datetime import datetime
-from pydantic import Field
-from pymongo import IndexModel, ASCENDING, DESCENDING
 from typing import List, Optional
 from uuid import UUID
 
-from depot_server.db.model.base import BaseDocument, BaseSubDocument
-from depot_server.model import ItemCondition, ReportState, TotalReportState
-from depot_server.helper.util import date
+from pydantic import Field
+from pymongo import IndexModel, ASCENDING, DESCENDING
 
+from depot_server.api.models import ItemCondition, ReportState, TotalReportState
+from depot_server.db.model.base import BaseDocument, BaseSubDocument
+from depot_server.helper.util import date
 
 
 class DbStrChange(BaseSubDocument):
