+++ src/depot_server/db/model/reservation.py
@@ -1,12 +1,12 @@
-from pydantic import Field
-from pymongo import IndexModel, ASCENDING
 from typing import Optional
 from uuid import UUID
 
-from depot_server.db.model.base import BaseDocument
-from depot_server.model import ReservationType, ReservationState
-from depot_server.helper.util import date
+from pydantic import Field
+from pymongo import IndexModel, ASCENDING
 
+from depot_server.api.models import ReservationType, ReservationState
+from depot_server.db.model.base import BaseDocument
+from depot_server.helper.util import date
 
 
 class DbReservation(BaseDocument):
