--- src/depot_server/model/__init__.py
@@ -1,11 +1,0 @@
-from .bay import Bay, BayInWrite
-from .device_reservation import DeviceReservation
-from .item import Item, ItemInWrite, ItemCondition, ReportItemInWrite
-from .item_state import ItemState, ItemStateChanges, StrChange, DateChange, ItemConditionChange, IdChange, \
-    ReportState, TotalReportStateChange
-from .picture import Picture
-from .report_element import ReportElement, ReportElementInWrite, ReportState
-from .report_profile import ReportProfile, ReportProfileInWrite, TotalReportState
-from .reservation import Reservation, ReservationInWrite, ReservationType, ReservationState, ReservationActionInWrite, \
-    ReservationAction, ReservationItemState, ReservationItem
-from .user import User
